chore: fixes cursor comment over circular dependencies on gas-speed and time.ts
