+++ app/components/Views/confirmations/hooks/gas/useGasFeeEstimateLevelOptions.ts
@@ -15,7 +15,6 @@ import { useFeeCalculations } from './useFeeCalculations';
 import { updateTransactionGasFees } from '../../../../../util/transaction-controller';
 import { type GasOption } from '../../types/gas';
 import { EMPTY_VALUE_STRING } from '../../constants/gas';
-import { isFastNetwork } from '../../components/gas/gas-speed/gas-speed';
 
 const HEX_ZERO = '0x0';
 
@@ -78,16 +77,11 @@ export const useGasFeeEstimateLevelOptions = ({
         }
       }
 
-      const estimatedTime = isFastNetwork(chainId)
-        ? toHumanEstimatedTimeRange(
-            networkGasFeeEstimates[level].minWaitTimeEstimate,
-            networkGasFeeEstimates[level].maxWaitTimeEstimate,
-            chainId,
-          )
-        : toHumanEstimatedTimeRange(
-            networkGasFeeEstimates[level].minWaitTimeEstimate,
-            networkGasFeeEstimates[level].maxWaitTimeEstimate,
-          );
+      const estimatedTime = toHumanEstimatedTimeRange(
+        networkGasFeeEstimates[level].minWaitTimeEstimate,
+        networkGasFeeEstimates[level].maxWaitTimeEstimate,
+        chainId,
+      );
 
       let feePerGas = HEX_ZERO;
       let gasPrice = HEX_ZERO;
